cncmark/pyproject.toml
[tool.poetry]
name = "cncmark"
version = "0.2.2"
description = ""
authors = ["yuichiroaoki <45054071+yuichiroaoki@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.17"
numpy-stl = "^3.0.1"
mysql-connector-python = "^8.1.0"
scipy = "^1.11.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"
ruff = "^0.0.282"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
